Move validation util functions
