+++ crates/sncast/tests/docs_snippets/validation.rs
@@ -1,6 +1,6 @@
-use std::collections::HashMap;
-use std::fs;
-
+use crate::helpers::devnet::{prepare_accounts_file, setup_contracts_map};
+use crate::helpers::fixtures::copy_directory_to_tempdir;
+use crate::helpers::runner::runner;
 use camino::Utf8PathBuf;
 use docs::snippet::{Snippet, SnippetType};
 use docs::utils::{
@@ -8,124 +8,7 @@ use docs::utils::{
     update_scarb_toml_dependencies,
 };
 use docs::validation::{extract_snippets_from_directory, extract_snippets_from_file};
-use regex::Regex;
-use shared::test_utils::output_assert::{assert_stdout_contains, AsOutput};
-use tempfile::TempDir;
-
-use crate::helpers::constants::URL;
-use crate::helpers::fixtures::copy_directory_to_tempdir;
-use crate::helpers::runner::runner;
-
-struct Contract {
-    class_hash: String,
-    contract_address: String,
-}
-
-fn prepare_accounts_file(temp: &TempDir) -> Utf8PathBuf {
-    // Account from predeployed accounts in starknet-devnet-rs
-    let accounts = r#"
-    {
-        "alpha-sepolia": {
-            "my_account": {
-            "address": "0x6f4621e7ad43707b3f69f9df49425c3d94fdc5ab2e444bfa0e7e4edeff7992d",
-            "deployed": true,
-            "private_key": "0x0000000000000000000000000000000056c12e097e49ea382ca8eadec0839401",
-            "public_key": "0x048234b9bc6c1e749f4b908d310d8c53dae6564110b05ccf79016dca8ce7dfac",
-            "type": "open_zeppelin"
-            }
-        }
-    }
-    "#;
-
-    let accounts_path = temp.path().join("accounts.json");
-    fs::write(&accounts_path, accounts).expect("Failed to write accounts.json");
-
-    Utf8PathBuf::from_path_buf(accounts_path).expect("Invalid UTF-8 path")
-}
-
-fn declare_and_deploy_contract(
-    contract_name: &str,
-    accounts_file: &str,
-    temp: &TempDir,
-) -> Contract {
-    let args = vec![
-        "--accounts-file",
-        accounts_file,
-        "--account",
-        "my_account",
-        "declare",
-        "--url",
-        URL,
-        "--contract-name",
-        contract_name,
-        "--max-fee",
-        "99999999999999999",
-        "--fee-token",
-        "strk",
-    ];
-
-    let snapbox = runner(&args).current_dir(temp.path());
-    let output = snapbox.assert().success();
-    let re_class_hash = Regex::new(r"class_hash:\s+(0x[a-fA-F0-9]+)").unwrap();
-
-    let class_hash = re_class_hash
-        .captures(output.as_stdout())
-        .and_then(|captures| captures.get(1))
-        .map(|match_| match_.as_str())
-        .expect("class_hash not found in the output");
-
-    let args = vec![
-        "--accounts-file",
-        accounts_file,
-        "--account",
-        "my_account",
-        "deploy",
-        "--url",
-        URL,
-        "--class-hash",
-        class_hash,
-        "--max-fee",
-        "99999999999999999",
-        "--fee-token",
-        "strk",
-    ];
-
-    let re_contract_address = Regex::new(r"contract_address:\s+(0x[a-fA-F0-9]+)").unwrap();
-
-    let snapbox = runner(&args).current_dir(temp.path());
-    let output = snapbox.assert().success();
-
-    let contract_address = re_contract_address
-        .captures(output.as_stdout())
-        .and_then(|captures| captures.get(1))
-        .map(|match_| match_.as_str())
-        .expect("contract_address not found in the output");
-
-    Contract {
-        class_hash: class_hash.to_string(),
-        contract_address: contract_address.to_string(),
-    }
-}
-
-fn setup_contracts_map(
-    tempdir: &TempDir,
-    account_json_path: &Utf8PathBuf,
-) -> HashMap<String, Contract> {
-    let mut contracts: HashMap<String, Contract> = HashMap::new();
-    let contract_names = [
-        "HelloSncast",
-        "DataTransformerContract",
-        "ConstructorContract",
-    ];
-
-    for contract_name in &contract_names {
-        let contract =
-            declare_and_deploy_contract(contract_name, account_json_path.as_str(), tempdir);
-        contracts.insert((*contract_name).to_string(), contract);
-    }
-
-    contracts
-}
+use shared::test_utils::output_assert::assert_stdout_contains;
 
 fn swap_next_element<'a>(args: &mut [&'a str], target: &str, new_value: &'a str) {
     if let Some(index) = args.iter().position(|&x| x == target) {
